Fetch GIDs for device nodes directly from stat instead of relying on `render` and `video` named groups (#2)
Came across a case in CI (GHA) where the self-hosted AMDGPU runner
(MI325) had an improperly configured `/dev/kfd` node, which wasn't owned
by the `render` group.

```
$ ls -ltra /dev/kfd
crw-rw---- 1 root  992 238, 0 Aug  5 11:11 kfd
```
Notice the owning group's ID is 992.

If setup properly, this device node needs to be owned by the `render`
group which already exists with a different GID:
```
$ getent group render
render:x:109:runner
```

This means even if I add the docker user (in this case `runner`) to the
render and video groups, I can't access the AMD GPUs in the container,
and I'd very much like to avoid hard-coding `--group-add 992` to the
docker run command for this case as that is too brittle.

This is a decent workaround as it directly gets the GID of the owning
group and passes it to the docker run cmd.