[Dev Containers] Support for sourcing env (PATH, LD_LIBRARY_PATH) via `activate`  (#6)
Adds an `activate` script that manages the user env within the container
just like a virtual environment. With this fix we should be able to
attach to an existing container without relying on the `entrypoint.sh`
to export the relevant env vars (`$PATH`, `$LD_LIBRARY_PATH`). It also
lets you `deactivate` to go back to the previous state of your shell.